Split 'routing' module into object-specific submodules
The 'routing' module became way too large and implements several almost
independent objects. This commit splits it into object-specific module
in 'netsim/modules/routing' directory:

* Common functionality (__init__.py)
* Shared normalize/import functions (normalize.py)
* AS path access list (aspath.py)
* Community lists (clist.py)
* Prefix lists (prefix.py)
* Routing policies (policy.py)
* Static routes (static.py)
