+++ netsim/modules/_routing.py
@@ -17,7 +17,7 @@ from ..data import global_vars
 from ..utils import log
 from ..utils import routing as _rp_utils
 from . import get_effective_module_attribute
-from .routing import check_routing_policy, import_routing_policy
+from .routing.policy import check_routing_policy, import_routing_policy
 
 
 # Build routing protocol address families
